services/transaction-service/pyproject.toml
[project]
name = "transaction-service"
version = "0.1.0"
description = ""
authors = [
    {name = "Kolade Fajimi",email = "fajimikolade23@gmail.com"}
]
requires-python = ">=3.11"
dependencies = [
    "grpcio (>=1.78.0,<2.0.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "aioboto3 (>=15.5.0,<16.0.0)",
    "redis (>=7.3.0,<8.0.0)",
    "protobuf (>=7.34.0,<8.0.0)",
    "pytest (>=9.0.2,<10.0.0)"
]


[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[dependency-groups]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "aiosqlite>=0.20",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
